Mern FullStack

+++ frontend/src/components/WorkoutDetails.jsx
@@ -2,14 +2,23 @@ import { useWorkOutContext } from "../hooks/useWorkoutContext";
 // date fns
 import formatDistanceToNow from "date-fns/formatDistanceToNow"
 
+import {useAuthContext } from '../hooks/useAuthContext'
+
 
 const WorkoutDetails = ({ title, load, reps, createdAt, _id }) => {
-  const { dispatch } = useWorkOutContext() ; // This line is added to use the dispatch function from the context.
+  const { dispatch } = useWorkOutContext() ;
+  const { user } = useAuthContext() // This line is added to use the dispatch function from the context.
   const httpUrl = `http://localhost:3000/api/workouts/${_id}`;
   const handleClick = async ()=>{
+    if(!user){
+      return
+    }
     try {
       const response = await fetch(httpUrl,{
-        method: 'DELETE'
+        method: 'DELETE',
+        headers : {
+          "Authorization": `Bearer ${user.token}`
+        }
       })
       const json  = await response.json()
       if (response.ok){
